Jenkinsfile_image_tester: rework the images naming
As it was done in Jenkinsfile_setup_new_machine remove the "bios"/"efi"
part in image name and the Yocto's DISTRO.

Allow the possibility to have a different image for each hypervisor
(for example, a hypervisor with a BIOS image and the other with an EFI
image).

Rename the NO_IOMMU option NO_DPDK because now the DISTRO part is no
longer present in the image name, only the using of DPDK or not differ.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>